High: accept descriptive source/output names
Instead of specifying raw pin numbers, the code now uses a lookup
table to define names (D3, A1 ...), and validate if these can be
used for input or output.

This is a preparation for being able to add I²C, SPI attached IOs that
could not be referenced via a single number.

Also, better error checking.